Order.API: configured API, established relations with Order->Basket and Basket->Product services.

+++ src/Order/Order.API/Models/Dto/OrderDto.cs
@@ -4,7 +4,7 @@ public class OrderDto
 {
     public int Id { get; set; }
     public string Login { get; set; } = null!;
-    public decimal TotalSum { get; set; }
+    public double TotalSum { get; set; }
     public DateTime DateOfOrder { get; set; }
     public ICollection<ShoppingCartDto> ShoppingCarts { get; set; } = new List<ShoppingCartDto>();
 }
